+++ src/components/User/Assets/AssetPairs/components/AssetPairManager.vue
@@ -31,41 +31,106 @@
 
     <template v-if="isLoaded">
       <form
-        @submit.prevent="updatePolicy"
+        @submit.prevent="updateAttributes"
         class="asset-pair-manager__form app__block"
       >
-        <h2>Update {{ base }}/{{ quote }} pair policies</h2>
+        <h2>Update {{ base }}/{{ quote }} pair attributes</h2>
+
+        <div class="asset-pair-manager__form-row app__form-row">
+          <input-field
+            class="app__form-field"
+            label="Physical price correction"
+            type="number"
+            min="0"
+            :step="DEFAULT_INPUT_STEP"
+            v-model="form.physicalPriceCorrection"
+            :disabled="isSubmitting"
+          >
+            <template slot="help">
+              <p class="asset-pair-manager__tip-message">
+                <span>
+                  The correction of physical price in percents.
+                  If physical price and restriction by physical price are set,
+                  minimum price for this pair offer will be
+                </span>
+                <strong>physicalPrice * physicalPriceCorrection</strong>
+              </p>
+            </template>
+          </input-field>
+        </div>
+
+        <div class="asset-pair-manager__form-row app__form-row">
+          <input-field
+            class="app__form-field"
+            label="Max price step"
+            type="number"
+            min="0"
+            :step="DEFAULT_INPUT_STEP"
+            v-model="form.maxPriceStep"
+            :disabled="isSubmitting"
+          >
+            <template slot="help">
+              <p class="asset-pair-manager__tip-message">
+                <span>
+                  Maximum offer price step in percents.
+                  If current price restriction is set,
+                  the users are allowed to set an offer with price in interval
+                </span>
+                <strong>(1 ± maxPriceStep) * currentPrice</strong>
+              </p>
+            </template>
+          </input-field>
+        </div>
 
         <div class="asset-pair-manager__checkboxes">
           <tick-field
             class="asset-pair-manager__checkbox"
             v-model="form.policies"
-            :disabled="isPending"
+            :disabled="isSubmitting"
             :required="false"
             label="Is tradable"
-            title="Allowed to trade this pair on secondary market"
             :cb-value="ASSET_PAIR_POLICIES.tradeableSecondaryMarket"
-          />
-          <!-- eslint-disable max-len -->
+          >
+            <template slot="help">
+              <span class="asset-pair-manager__tip-message">
+                Allowed to trade this pair on secondary market
+              </span>
+            </template>
+          </tick-field>
+
           <tick-field
             class="asset-pair-manager__checkbox"
             v-model="form.policies"
-            :disabled="isPending"
+            :disabled="isSubmitting"
             :required="false"
             label="Physical price restriction"
-            title="If set, then prices for new offers must be greater then physical price with correction"
             :cb-value="ASSET_PAIR_POLICIES.physicalPriceRestriction"
-          />
+          >
+            <template slot="help">
+              <span class="asset-pair-manager__tip-message">
+                If set, then prices for new offers must be greater
+                than physical price with correction
+              </span>
+            </template>
+          </tick-field>
+
           <tick-field
             class="asset-pair-manager__checkbox"
             v-model="form.policies"
-            :disabled="isPending"
+            :disabled="isSubmitting"
             :required="false"
             label="Current price restriction"
-            title="If set, then price for new offers must be in interval of (1 +- maxPriceStep)*currentPrice"
             :cb-value="ASSET_PAIR_POLICIES.currentPriceRestriction"
-          />
-          <!-- eslint-enable max-len -->
+          >
+            <template slot="help">
+              <p class="asset-pair-manager__tip-message">
+                <span>
+                  If set, then price for new offers must be in interval of
+                </span>
+                <strong>(1 ± maxPriceStep) * currentPrice</strong>
+              </p>
+            </template>
+          </tick-field>
         </div>
 
         <button
@@ -117,6 +182,8 @@ export default {
       pair: {},
       form: {
         price: '',
+        physicalPriceCorrection: '',
+        maxPriceStep: '',
         policies: [],
       },
       isLoaded: false,
@@ -127,6 +194,8 @@ export default {
   async created () {
     await this.getPair()
     this.form.price = this.pair.physicalPrice
+    this.form.physicalPriceCorrection = this.pair.physicalPriceCorrection
+    this.form.maxPriceStep = this.pair.maxPriceStep
     this.form.policies = this.pair.policies.map(policy => policy.value)
   },
 
@@ -152,7 +221,10 @@ export default {
       this.submit({ updatePrice: true })
     },
 
-    async updatePolicy () {
+    async updateAttributes () {
+      this.pair.physicalPriceCorrection = this.form.physicalPriceCorrection
+      this.pair.maxPriceStep = this.form.maxPriceStep
+
       this.pair.policies = this.form.policies
         .reduce((sum, policy) => sum | policy, 0)
       this.submit({ updatePolicy: true })
@@ -177,16 +249,23 @@ export default {
 <style scoped>
 .asset-pair-manager {
   display: flex;
+  justify-content: center;
 }
 
 .asset-pair-manager__form {
   margin-right: 30px;
+  height: 100%;
 }
 
 .asset-pair-manager__checkboxes {
-  margin: 6.4rem 0 4rem;
+  margin: 4.5rem 0 4rem;
 }
+
 .asset-pair-manager__checkbox {
   margin: 0 0 1rem;
+}
+
+.asset-pair-manager__tip-message {
+  min-width: 20rem;
 }
 </style>