+++ src/components/common/fields/InputField.vue
@@ -3,6 +3,16 @@
     class="input-field"
     :class="{'input-field--error': errorMessage}"
   >
+    <div
+      v-if="$slots.help"
+      class="input-field__tip"
+    >
+      <mdi-help-circle-icon class="input-field__tip-icon" />
+      <div class="input-field__tip-content">
+        <slot name="help" />
+      </div>
+    </div>
+
     <input
       class="input-field__input"
       :class="{ 'input-field__input--placeholder-auto-hidden': label }"
@@ -47,6 +57,8 @@
 
 <script>
 import InputFieldAutocomplete from './InputFieldAutocomplete'
+
+import 'mdi-vue/HelpCircleIcon'
 
 export default {
   components: {